Move internal rule model related code to sub-folder model
