+++ workspace/src/app/views/shared/RuleEditor/model/RuleEditorModel.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 import { Edge, Elements, OnLoadParams, removeElements, useStoreActions } from "react-flow-renderer";
-import { RuleEditorModelContext } from "./contexts/RuleEditorModelContext";
-import { RuleEditorContext, RuleEditorContextProps } from "./contexts/RuleEditorContext";
+import { RuleEditorModelContext } from "../contexts/RuleEditorModelContext";
+import { RuleEditorContext, RuleEditorContextProps } from "../contexts/RuleEditorContext";
 import utils from "./RuleEditorModel.utils";
 import { useTranslation } from "react-i18next";
-import { IRuleOperator, IRuleOperatorNode } from "./RuleEditor.typings";
+import { IRuleOperator, IRuleOperatorNode } from "../RuleEditor.typings";
 import {
     AddEdge,
     AddNode,
